pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yolov8-csv-to-yolo"
version = "1.0.0"
description = "Professional YOLOv8 object detection training pipeline with CSV annotation support"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [{name = "YOLOv8 Pipeline Team"}]
keywords = ["yolov8", "object-detection", "computer-vision", "deep-learning", "csv", "annotation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
]
dependencies = [
    "ultralytics>=8.2",
    "opencv-python-headless>=4.8",
    "pandas>=2.0",
    "pyyaml>=6.0",
    "matplotlib>=3.7",
    "Pillow>=10.0",
    "tqdm>=4.65",
]

[project.optional-dependencies]
export = ["onnx>=1.14", "onnxruntime>=1.16"]
dev = ["pytest", "black", "ruff", "pre-commit"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
